[FIX] runbot: speedup and limit search in frontend
When searching the builds for the frontend the resulting query can last
a very long time (up to 7sec).

With this commit, the search result is strictly limited to 1000 builds
and the search string length is limited to 60 chars.

Also, the branch_id many2one field on build is set to auto_join in order
to generate a better sql query. Thanks @tbe-odoo.