[REF] runbot: split cron to speed up builds
Then Runbot Cron is executed on each runbot instance. When the number of
instances scales, the time needed for an instance to obtain the cron
increases.

With this commit, the Runbot cron only executes the _update mtehod.
That way, only one runbot instance fetch the respositories and create
pending builds.

On the other hand, a dedicated cron have to be manually created for each
runbot instance. Those cron's only have to call the _cron_by_host
method, passing the hostname as parameter. This method will then self
assign pending builds if there are slots available. In that case, if the
build hash cannot be found in the local repo, the repo is fetched.